requirements.txt
Django==3.2.9
Pillow==8.4.0
django-cleanup==6.0.0
python-decouple==3.5
Django==3.2.9
python-decouple==3.5
whitenoise==5.3.0
boto3==1.20.3
botocore==1.23.23
asgiref==3.4.1
gunicorn==20.1.0
urllib3==1.26.7
six==1.16.0
django-storages